third_party/WebKit/LayoutTests/editing/deleting/delete-inline-br-expected.txt

PASS caretRect.left is expectedCaretRect.left
PASS caretRect.top is expectedCaretRect.top
